Add pytest command line option to skip resource teardown (#3291)

+++ tests/suite/fixtures/custom_resource_fixtures.py
@@ -82,10 +82,11 @@ def virtual_server_setup(request, kube_apis, ingress_controller_endpoint, test_n
         wait_until_all_pods_are_ready(kube_apis.v1, test_namespace)
 
     def fin():
-        print("Clean up Virtual Server Example:")
-        delete_virtual_server(kube_apis.custom_objects, vs_name, test_namespace)
-        if request.param.get("app_type"):
-            delete_common_app(kube_apis, request.param["app_type"], test_namespace)
+        if request.config.getoption("--skip-fixture-teardown") == "no":
+            print("Clean up Virtual Server Example:")
+            delete_virtual_server(kube_apis.custom_objects, vs_name, test_namespace)
+            if request.param.get("app_type"):
+                delete_common_app(kube_apis, request.param["app_type"], test_namespace)
 
     request.addfinalizer(fin)
 
@@ -141,10 +142,11 @@ def transport_server_setup(
     wait_until_all_pods_are_ready(kube_apis.v1, test_namespace)
 
     def fin():
-        print("Clean up TransportServer Example:")
-        delete_ts(kube_apis.custom_objects, ts_resource, test_namespace)
-        delete_items_from_yaml(kube_apis, service_file, test_namespace)
-        delete_gc(kube_apis.custom_objects, gc_resource, "nginx-ingress")
+        if request.config.getoption("--skip-fixture-teardown") == "no":
+            print("Clean up TransportServer Example:")
+            delete_ts(kube_apis.custom_objects, ts_resource, test_namespace)
+            delete_items_from_yaml(kube_apis, service_file, test_namespace)
+            delete_gc(kube_apis.custom_objects, gc_resource, "nginx-ingress")
 
     request.addfinalizer(fin)
 
@@ -186,13 +188,14 @@ def v_s_route_app_setup(request, kube_apis, v_s_route_setup) -> None:
     wait_until_all_pods_are_ready(kube_apis.v1, v_s_route_setup.route_s.namespace)
 
     def fin():
-        print("Clean up the Application:")
-        delete_deployment(kube_apis.apps_v1_api, deployment_one, v_s_route_setup.route_m.namespace)
-        delete_service(kube_apis.v1, svc_one, v_s_route_setup.route_m.namespace)
-        delete_deployment(kube_apis.apps_v1_api, deployment_three, v_s_route_setup.route_m.namespace)
-        delete_service(kube_apis.v1, svc_three, v_s_route_setup.route_m.namespace)
-        delete_deployment(kube_apis.apps_v1_api, deployment_two, v_s_route_setup.route_s.namespace)
-        delete_service(kube_apis.v1, svc_two, v_s_route_setup.route_s.namespace)
+        if request.config.getoption("--skip-fixture-teardown") == "no":
+            print("Clean up the Application:")
+            delete_deployment(kube_apis.apps_v1_api, deployment_one, v_s_route_setup.route_m.namespace)
+            delete_service(kube_apis.v1, svc_one, v_s_route_setup.route_m.namespace)
+            delete_deployment(kube_apis.apps_v1_api, deployment_three, v_s_route_setup.route_m.namespace)
+            delete_service(kube_apis.v1, svc_three, v_s_route_setup.route_m.namespace)
+            delete_deployment(kube_apis.apps_v1_api, deployment_two, v_s_route_setup.route_s.namespace)
+            delete_service(kube_apis.v1, svc_two, v_s_route_setup.route_s.namespace)
 
     request.addfinalizer(fin)
 
@@ -288,14 +291,15 @@ def v_s_route_setup(request, kube_apis, ingress_controller_endpoint) -> VirtualS
     route_s = VirtualServerRoute(ns_2, vsr_s_name, vsr_s_paths)
 
     def fin():
-        print("Clean up the Virtual Server Route:")
-        delete_v_s_route(kube_apis.custom_objects, vsr_m_name, ns_1)
-        delete_v_s_route(kube_apis.custom_objects, vsr_s_name, ns_2)
-        print("Clean up Virtual Server:")
-        delete_virtual_server(kube_apis.custom_objects, vs_name, ns_1)
-        print("Delete test namespaces")
-        delete_namespace(kube_apis.v1, ns_1)
-        delete_namespace(kube_apis.v1, ns_2)
+        if request.config.getoption("--skip-fixture-teardown") == "no":
+            print("Clean up the Virtual Server Route:")
+            delete_v_s_route(kube_apis.custom_objects, vsr_m_name, ns_1)
+            delete_v_s_route(kube_apis.custom_objects, vsr_s_name, ns_2)
+            print("Clean up Virtual Server:")
+            delete_virtual_server(kube_apis.custom_objects, vs_name, ns_1)
+            print("Delete test namespaces")
+            delete_namespace(kube_apis.v1, ns_1)
+            delete_namespace(kube_apis.v1, ns_2)
 
     request.addfinalizer(fin)
