+++ tests/suite/test_app_protect_integration.py
@@ -99,14 +99,15 @@ def appprotect_setup(request, kube_apis, ingress_controller_endpoint, test_names
     create_items_from_yaml(kube_apis, src_syslog_yaml, test_namespace)
 
     def fin():
-        print("Clean up:")
-        delete_items_from_yaml(kube_apis, src_syslog_yaml, test_namespace)
-        delete_ap_policy(kube_apis.custom_objects, pol_name, test_namespace)
-        delete_ap_logconf(kube_apis.custom_objects, log_name, test_namespace)
-        delete_common_app(kube_apis, "simple", test_namespace)
-        src_sec_yaml = f"{TEST_DATA}/appprotect/appprotect-secret.yaml"
-        delete_items_from_yaml(kube_apis, src_sec_yaml, test_namespace)
-        write_to_json(f"reload-{get_test_file_name(request.node.fspath)}.json", reload_times)
+        if request.config.getoption("--skip-fixture-teardown") == "no":
+            print("Clean up:")
+            delete_items_from_yaml(kube_apis, src_syslog_yaml, test_namespace)
+            delete_ap_policy(kube_apis.custom_objects, pol_name, test_namespace)
+            delete_ap_logconf(kube_apis.custom_objects, log_name, test_namespace)
+            delete_common_app(kube_apis, "simple", test_namespace)
+            src_sec_yaml = f"{TEST_DATA}/appprotect/appprotect-secret.yaml"
+            delete_items_from_yaml(kube_apis, src_sec_yaml, test_namespace)
+            write_to_json(f"reload-{get_test_file_name(request.node.fspath)}.json", reload_times)
 
     request.addfinalizer(fin)
