+++ tests/suite/test_auth_basic_secrets.py
@@ -72,11 +72,14 @@ def auth_basic_secrets_setup(
     )
 
     def fin():
-        print("Clean up the Auth Basic Secrets Application:")
-        delete_common_app(kube_apis, "simple", test_namespace)
-        delete_items_from_yaml(
-            kube_apis, f"{TEST_DATA}/auth-basic-secrets/{request.param}/auth-basic-secrets-ingress.yaml", test_namespace
-        )
+        if request.config.getoption("--skip-fixture-teardown") == "no":
+            print("Clean up the Auth Basic Secrets Application:")
+            delete_common_app(kube_apis, "simple", test_namespace)
+            delete_items_from_yaml(
+                kube_apis,
+                f"{TEST_DATA}/auth-basic-secrets/{request.param}/auth-basic-secrets-ingress.yaml",
+                test_namespace,
+            )
 
     request.addfinalizer(fin)
 
@@ -93,9 +96,10 @@ def auth_basic_secret(
     wait_before_test(1)
 
     def fin():
-        print("Delete Secret:")
-        if is_secret_present(kube_apis.v1, secret_name, test_namespace):
-            delete_secret(kube_apis.v1, secret_name, test_namespace)
+        if request.config.getoption("--skip-fixture-teardown") == "no":
+            print("Delete Secret:")
+            if is_secret_present(kube_apis.v1, secret_name, test_namespace):
+                delete_secret(kube_apis.v1, secret_name, test_namespace)
 
     request.addfinalizer(fin)
